Fundamental Kotlin - Classes.
